dooshki_args: Add support for the double-dash "--" command-line keyword.
In case the user wishes to pass a list of arguments to a program
which should not be touched by the option parsing code, they can
pass them after a "--" keyword.  Any argument after the keyword
will be ignored and left untouched in the arguments list (in argv).

For example, a wrapper for a separate executable might accept
command-line options for itself, as well as for the underlying
executable, and these two sets of command-line options can be
separated using the "--" keyword.