dooshki_args: Make short option handling more similar to POSIX.
The previous implementation of argument handling for short options
turned out to cause confusion for users used to specifying arguments
without a space, as is permitted by POSIX.

This change allows the arguments of short options to be specified
right after the option letter itself, either with a space, an equals
sign, or with no space at all.  It removes the ability to specify
multiple short options that accept an argument within a single
short argument string, as the first short option that accepts an
argument immediately consumes the rest of the string.